requirements.txt
# 核心依赖
torch>=2.0.0
torchvision>=0.15.0

# Hugging Face 库
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.12.0
huggingface-hub>=0.16.0

# 图像处理
pillow>=9.5.0
opencv-python>=4.7.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0

# 配置和工具
pyyaml>=6.0
tqdm>=4.65.0
tensorboard>=2.13.0

# 其他工具
einops>=0.6.1
safetensors>=0.3.1
matplotlib>=3.7.0

